+++ include/crypto/ml_kem.h
@@ -122,49 +122,8 @@
  * -----------------------------
  */
 
- /*-
-  * The DER ASN.1 encoding of ML-KEM (and ML-DSA) public keys prepends 22 bytes
-  * to the encoded public key:
-  *
-  * - 4 byte outer sequence tag and length
-  * -  2 byte algorithm sequence tag and length
-  * -    2 byte algorithm OID tag and length
-  * -      9 byte algorithm OID (from NIST CSOR OID arc)
-  * -  4 byte bit string tag and length
-  * -    1 bitstring lead byte
-  */
-# define ML_KEM_SPKI_OVERHEAD   22
-typedef struct {
-    const uint8_t asn1_prefix[ML_KEM_SPKI_OVERHEAD];
-} ML_KEM_SPKI_INFO;
-
-/*-
- * For each algorithm we support a few PKCS#8 input formats,
- *
- * - Seed: SEQUENCE(OCTET STRING)
- * - Private key: SEQUENCE([1] IMPLICIT OCTET STRING)
- * - Seed & private key: SEQUENCE(OCTET STRING, [1] IMPLICIT OCTET STRING)
- * - OQS private key: OCTET STRING
- * - OQS private + public key: OCTET STRING
- *   (The public key is ignored, just as with PKCS#8 v2.)
- *
- * An offset of zero means that particular field is absent.
- */
-typedef struct {
-    const char *p8_name;
-    size_t p8_bytes;
-    uint32_t p8_magic;
-    uint16_t seed_magic;
-    size_t seed_offset;
-    uint32_t priv_magic;
-    size_t priv_offset;
-    size_t pub_offset;
-} ML_KEM_PKCS8_INFO;
-
 typedef struct {
     const char *algorithm_name;
-    const ML_KEM_SPKI_INFO *spki_info;
-    const ML_KEM_PKCS8_INFO *pkcs8_info;
     size_t prvkey_bytes;
     size_t prvalloc;
     size_t pubkey_bytes;
@@ -212,6 +171,7 @@ typedef struct ossl_ml_kem_key_st {
     struct ossl_ml_kem_scalar_st *s;        /* Private key secret vector */
     uint8_t *z;                             /* Private key FO failure secret */
     uint8_t *d;                             /* Private key seed */
+    int prefer_seed;                        /* Given seed and key use seed? */
     int retain_seed;                        /* Retain the seed after keygen? */
 
     /*
@@ -221,13 +181,17 @@ typedef struct ossl_ml_kem_key_st {
      * |z| and |d| components in that order.
      */
     uint8_t seedbuf[64];                    /* |rho| + |pkhash| / |z| + |d| */
+    uint8_t *encoded_dk;                    /* Unparsed P8 private key */
 } ML_KEM_KEY;
 
 /* The public key is always present, when the private is */
-# define ossl_ml_kem_key_vinfo(key)         ((key)->vinfo)
-# define ossl_ml_kem_have_pubkey(key)       ((key)->t != NULL)
-# define ossl_ml_kem_have_prvkey(key)       ((key)->s != NULL)
-# define ossl_ml_kem_have_seed(key)         ((key)->d != NULL)
+# define ossl_ml_kem_key_vinfo(key)     ((key)->vinfo)
+# define ossl_ml_kem_have_pubkey(key)   ((key)->t != NULL)
+# define ossl_ml_kem_have_prvkey(key)   ((key)->s != NULL)
+# define ossl_ml_kem_have_seed(key)     ((key)->d != NULL)
+# define ossl_ml_kem_have_dkenc(key)    ((key)->encoded_dk != NULL)
+# define ossl_ml_kem_decoded_key(key)   ((key)->encoded_dk != NULL \
+                                         || ((key)->s == NULL && (key)->d != NULL))
 
 /*
  * ----- ML-KEM key lifecycle
@@ -237,9 +201,8 @@ typedef struct ossl_ml_kem_key_st {
  * Allocate a "bare" key for given ML-KEM variant. Initially without any public
  * or private key material.
  */
-ML_KEM_KEY *ossl_ml_kem_key_new(OSSL_LIB_CTX *libctx,
-                                const char *properties,
-                                int retain_seed, int evp_type);
+ML_KEM_KEY *ossl_ml_kem_key_new(OSSL_LIB_CTX *libctx, const char *properties,
+                                int evp_type);
 /* Deallocate the key */
 void ossl_ml_kem_key_free(ML_KEM_KEY *key);
 /*
@@ -268,17 +231,6 @@ int ossl_ml_kem_parse_private_key(const uint8_t *in, size_t len,
 ML_KEM_KEY *ossl_ml_kem_set_seed(const uint8_t *seed, size_t seedlen,
                                  ML_KEM_KEY *key);
 __owur
-ML_KEM_KEY *ossl_ml_kem_d2i_PUBKEY(const uint8_t *pubenc, int publen,
-                                   int evp_type, OSSL_LIB_CTX *libctx,
-                                   const char *propq);
-__owur
-ML_KEM_KEY *ossl_ml_kem_d2i_PKCS8(const uint8_t *prvenc, int prvlen,
-                                  int retain_seed, const char *formats,
-                                  int evp_type, OSSL_LIB_CTX *ctx,
-                                  const char *propq);
-__owur
-int ossl_ml_kem_key_to_text(BIO *out, const ML_KEM_KEY *key, int selection);
-__owur
 int ossl_ml_kem_genkey(uint8_t *pubenc, size_t publen, ML_KEM_KEY *key);
 
 /*
@@ -290,13 +242,9 @@ __owur
 int ossl_ml_kem_encode_public_key(uint8_t *out, size_t len,
                                   const ML_KEM_KEY *key);
 __owur
-int ossl_ml_kem_i2d_pubkey(const ML_KEM_KEY *key, unsigned char **out);
-__owur
 int ossl_ml_kem_encode_private_key(uint8_t *out, size_t len,
                                    const ML_KEM_KEY *key);
 __owur
-int ossl_ml_kem_i2d_prvkey(const ML_KEM_KEY *key, unsigned char **out,
-                           const char *formats);
 int ossl_ml_kem_encode_seed(uint8_t *out, size_t len,
                             const ML_KEM_KEY *key);
