+++ providers/implementations/encode_decode/build.info
@@ -18,3 +18,7 @@ IF[{- !$disabled{ec} -}]
   SOURCE[$ENCODER_GOAL]=encode_key2blob.c
 ENDIF
 DEPEND[encode_key2any.o]=../../common/include/prov/der_rsa.h
+
+IF[{- !$disabled{'ml-kem'} -}]
+  SOURCE[$DECODER_GOAL]=ml_kem_codecs.c
+ENDIF
