+++ providers/implementations/encode_decode/encode_key2text.c
@@ -29,6 +29,7 @@
 #include "prov/implementations.h"
 #include "internal/encoder.h"
 #include "endecoder_local.h"
+#include "ml_kem_codecs.h"
 
 DEFINE_SPECIAL_STACK_OF_CONST(BIGNUM_const, BIGNUM)
